% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algiers.R
\docType{data}
\name{Algiers}
\alias{Algiers}
\title{Daily maximum (tX) and minimum (tN) air temperatures for Algiers, Algeria.}
\format{
A data frame with 16436 rows and 3 variables:
\describe{
  \item{t}{date, as.Date() format}
  \item{tMax}{daily max. temperature, in degrees Celsius}
  \item{tMin}{daily min. temperature, in degrees Celsius}
  ...
}
}
\source{
Mr. Haouari Mahmoud, IHFR, Algeria
}
\usage{
Algiers
}
\description{
A dataset containing the daily maximum and minimum air temperatures
(in degrees Celsius) and date for Algiers, Algeria
for the period 1961-01-01 to 2005-12-31.
}
\details{
lon/lat:
}
\keyword{datasets}
