% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_calc.R
\name{clim_calc}
\alias{clim_calc}
\title{Calculate seasonal and threshold climatologies as wellas the varaince}
\usage{
clim_calc(data, window_half_width)
}
\arguments{
\item{data}{The data given to this function during the calculations
performed by \code{\link{detect_clim}}.}

\item{window_half_width}{The width of the smoothing window to be applied.
This width is doubled and centred around the point that the smoothing
occurs. Default = 5, which makes an overall window size of 11.}
}
\value{
The function returns the calculated climatologies.
}
\description{
An internal function that helps to create the climatologies that are
then output with \code{\link{detect_clim}}.
}
