% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_clim.R, R/detect_event.R
\name{detect_clim}
\alias{detect_clim}
\alias{detect_clim}
\title{Detect the climatology for a time series}
\usage{
detect_clim(data, x = t, y = temp, min_duration = 5,
  join_across_gaps = TRUE, max_gap = 2)

detect_clim(data, x = t, y = temp, min_duration = 5,
  join_across_gaps = TRUE, max_gap = 2)
}
\arguments{
\item{data}{A data frame with columns for date and temperature data.
Ordered daily data are expected, and although missing values (NA) can be
accommodated, the function is only recommended when NAs occur infrequently,
preferably at no more than 3 consecutive days.}

\item{x}{A column with the daily time vector (see details). For backwards
compatibility, the column is named \code{t} by default.}

\item{y}{A column with the response vector. RmarineHeatWaves version <= 0.15.9
assumed that this would be daily seawater temperatures, but as of version 0.16.0
it may be any arbitrary measurement taken at a daily frequency. The default
remains temperature, and the default column name is therefore \code{temp}, again
hopefully ensuring backwards compatibility.}

\item{min_duration}{Minimum duration for acceptance of detected MHWs.
Default is \code{5} days.}

\item{join_across_gaps}{Boolean switch indicating whether to join MHWs which
occur before/after a short gap as specified by \code{max_gap}. Default
is \code{TRUE}.}

\item{max_gap}{Maximum length of gap allowed for the joining of MHWs. Default
is \code{2} days.}

\item{climatology_start}{Required. The start date for the period across which
the (varying by day-of-year) seasonal cycle and extremes threshold are
calculated.}

\item{climatology_end}{Required. The end date for the period across which
the (varying by day-of-year) seasonal cycle and extremes threshold are
calculated.}

\item{pctile}{Threshold percentile (\%) for detection of extreme values.
Default is \code{90}th percentile. Please see \code{cold_spells} for more
information about the calculation of marine cold-spells.}

\item{window_half_width}{Width of sliding window about day-of-year (to one
side of the center day-of-year) used for the pooling of values and
calculation of climatology and threshold percentile. Default is \code{5}
days, which gives a window width of 11 days centered on the 6th day of the
series of 11 days.}

\item{smooth_percentile}{Boolean switch selecting whether to smooth the
climatology and threshold percentile timeseries with a moving average of
\code{smooth_percentile_width}. Default is \code{TRUE}.}

\item{smooth_percentile_width}{Full width of moving average window for smoothing
climatology and threshold. Default is \code{31} days.}

\item{max_pad_length}{Specifies the maximum length of days over which to
interpolate (pad) missing data (specified as \code{NA}) in the input
temperature time series; i.e., any consecutive blocks of NAs with length
greater than \code{max_pad_length} will be left as \code{NA}. Set as an
integer. The default is \code{3} days.}

\item{data}{A data frame with three columns. In the default setting (i.e. ommitting
the arguments \code{doy}, \code{x} and \code{y}; see immediately below), the
data set is expected to have the headers \code{doy}, \code{t} and \code{temp}.
\code{doy} is the Julian day running from 1 to 366, but modified so that the
day-of-year (doy) vector for non-leap-years runs 1...59 and then 61...366.
For leap years the 60th day is February 29. The \code{t} column is a vector
of dates of class \code{Date}, while \code{temp} is the measured variable (by
default it is assumed to be temperature). Data of the appropriate format are
created by the function \code{\link{detect_clim}}, but your own data can be supplied
if they meet the criteria specified by \code{\link{detect_clim}}.}

\item{doy}{If a column headed \code{doy} is not available, another column with
Julian dates can be supplied. This argument accepts the name of that column. The
default name is, of course, \code{doy}.}

\item{x}{This column is expected to contain a vector of dates as per the
specification of \code{make_whole}. If a column headed \code{t} is present in
the dataframe, this argument may be ommitted; otherwise, specify the name of
the column with dates here.}

\item{y}{This is a column containing the measurement variable. If the column
name differs from the default (i.e. \code{temp}), specify the name here.}
}
\description{
Applies a climatology detecting algorithm to a time series with
several built-in options. The output of this function may then be
used directly with the \code{\link{detect_event}} function.

Applies the Hobday et al. (2016) marine heat wave definition to an input time
series of a given value (temperature) along with a daily date vector.
}
