% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatwaveR-package.R
\docType{package}
\name{heatwaveR-package}
\alias{heatwaveR}
\alias{heatwaveR-package}
\title{heatwaveR: Detect heatwaves and cold-spells}
\description{
This package is designed to include and encompass the different methods of defining and detecting extreme events, known as heatwaves or cold-spells in both air and water temperature data. These detection algorithms may be used on non-temperature data as well however, this is not catered for explicitly here as no use of this technique in the literature currently exists.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://robwschlegel.github.io/heatwaveR/index.html}
}

}
\author{
\strong{Maintainer}: Robert W. Schlegel \email{robwschlegel@gmail.com} (0000-0002-0705-1287) [contributor]

Authors:
\itemize{
  \item Eric C. J. Oliver (The brain behind the Python implementation.)
  \item Albertus J. Smit (R implementation.) [contributor]
}

}
\keyword{internal}
