% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_wa.R
\docType{data}
\name{sst_WA}
\alias{sst_WA}
\title{NOAA Optimally Interpolated (OI) v2.1 daily 1/4 degree SST for the Western Australian region.}
\format{
A dataframe with 13879 rows and 2 variables:
\describe{
  \item{t}{date, as.Date() format}
  \item{temp}{SST, in degrees Celsius}
  ...
}
}
\source{
\url{https://www.ncdc.noaa.gov/oisst}
}
\usage{
sst_WA
}
\description{
A dataset containing the sea surface temperature (in degrees Celsius) and
date for the Western Australian region from 1982-01-01 to 2019-12-31.
}
\details{
lon/lat: 112.5/-29.5
}
\keyword{datasets}
