% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_calc.R
\name{clim_calc}
\alias{clim_calc}
\title{Calculate daily mean, threshold, and variance climatologies.}
\usage{
clim_calc(data, windowHalfWidth, pctile)
}
\arguments{
\item{data}{The data given to this function during the calculations
performed by \code{\link{ts2clm}}.}

\item{windowHalfWidth}{The width of the smoothing window to be applied.
This width is doubled and centred around the point that the smoothing
occurs. Default = 5, which makes an overall window size of 11.}

\item{pctile}{Threshold percentile (\%) for detection of events (MHWs).
Default is \code{90}th percentile.}
}
\value{
The function returns the calculated climatologies.
}
\description{
An internal function that helps to create the climatologies that are
then output with \code{\link{ts2clm}}.
}
\author{
Smit, A. J.
}
\keyword{internal}
