% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatwaveR-package.R
\docType{package}
\name{heatwaveR-package}
\alias{heatwaveR}
\alias{heatwaveR-package}
\title{heatwaveR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
