% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatwaveR-package.R
\docType{package}
\name{heatwaveR-package}
\alias{heatwaveR}
\alias{heatwaveR-package}
\title{heatwaveR: Detect Heatwaves and Cold-Spells}
\description{
The different methods of defining and detecting extreme events, known as heatwaves or cold-spells in both air and water temperature data are encompassed within this package. These detection algorithms may be used on non-temperature data as well however, this is not catered for explicitly here as no use of this technique in the literature currently exists.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://robwschlegel.github.io/heatwaveR/index.html}
  \item \url{https://github.com/robwschlegel/heatwaveR}
  \item Report bugs at \url{https://github.com/robwschlegel/heatwaveR/issues}
}

}
\author{
\strong{Maintainer}: Robert W. Schlegel \email{robwschlegel@gmail.com} (\href{https://orcid.org/0000-0002-0705-1287}{ORCID}) [contributor]

Authors:
\itemize{
  \item Albertus J. Smit (\href{https://orcid.org/0000-0002-3799-6126}{ORCID}) [contributor]
}

}
\keyword{internal}
