% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proto_event.R
\name{proto_event}
\alias{proto_event}
\title{Detect proto-events based on a chosen criterion (column).}
\usage{
proto_event(t_series, criterion_column, minDuration, gaps = FALSE, maxGap)
}
\arguments{
\item{t_series}{A dataframe of the correct dimensions inherited
from \code{\link{detect_event}} within which this runs.}

\item{criterion_column}{The column to use for the detection of events.}

\item{minDuration}{Minimum duration for acceptance of detected events.}

\item{gaps}{This logic gate tells this internal function if it should be
calculating the lengths of events, or rather the gaps in between them.
The default is \code{FALSE}.}

\item{maxGap}{This is the number of rows (days) across which distinct
events will be combined into one event if \code{joinAcrossGaps = TRUE}.}
}
\value{
A dataframe that will be used within
\code{\link{detect_event}}.
}
\description{
An internal function that detects proto-events, which are periods of
time above a threshold, but without necessarily considering a minimum
duration or joining across gaps.
}
\author{
Smit, A. J.
}
