% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_percentile.R
\name{smooth_percentile}
\alias{smooth_percentile}
\title{Detect the climatology for a time series.}
\usage{
smooth_percentile(data, smoothPercentileWidth)
}
\arguments{
\item{data}{The data given to this function during the calculations
performed by \code{\link{ts2clm}}.}

\item{smoothPercentileWidth}{The width of the smoothing window
to be applied. The default is \code{31} days.}
}
\value{
The function returns the data in the same format it was
input as, with the climatology values smoothed as desired.
}
\description{
An internal function that helps to create climatologies in
accordance with the Hobday et al. (2016) standard.
}
\author{
Smit, A. J.
}
