% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_med.R
\docType{data}
\name{sst_Med}
\alias{sst_Med}
\title{Optimally interpolated 0.25 degree SST for the Mediterranean region.}
\format{A data frame with 12053 rows and 2 variables:
\describe{
  \item{t}{date, as POSIXct}
  \item{temp}{SST, in degrees Celsius}
  ...
}}
\source{
\url{https://www.ncdc.noaa.gov/oisst}
}
\usage{
sst_Med
}
\description{
A dataset containing the sea surface temperature (in degrees Celsius)
and date for the Mediterranean region from 1982-01-01 to 2014-12-31.
}
\details{
lon/lat: 9/43.5
}
\keyword{datasets}
