% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_nw_atl.R
\docType{data}
\name{sst_NW_Atl}
\alias{sst_NW_Atl}
\title{Optimally interpolated 0.25 degree SST for the NW Atlantic region.}
\format{A data frame with 12053 rows and 2 variables:
\describe{
  \item{t}{date, as POSIXct}
  \item{temp}{SST, in degrees Celsius}
  ...
}}
\source{
\url{https://www.ncdc.noaa.gov/oisst}
}
\usage{
sst_NW_Atl
}
\description{
A dataset containing the sea surface temperature (in degrees Celsius) and
date for the Northwest Atlantic region from 1982-01-01 to 2014-12-31.
}
\details{
lon/lat: -67/43
}
\keyword{datasets}
